requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
cachelib==0.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cs50==5.1.0
Flask==1.1.2
Flask-Session==0.3.2
futures==3.1.1
goslate==1.5.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
nose==1.3.7
protobuf==3.13.0
PyDictionary==2.0.1
Random-Word==1.0.4
RandomWords==0.3.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
termcolor==1.1.0
urllib3==1.25.11
Werkzeug==1.0.1
